actualizacion de ejercicios
