--- EjerciciosJava/README.md
@@ -1,1 +1,0 @@
-proyecto que contiene ejercicios y clases de Java
